Support priority adjustments based on build trigger
As recommended by Martin Franklin, we can now adjust the priority of a job
based on how it was started.  The main use-case for this is to have builds
started by users run before build triggered from a timer or SCM change.  In
most cases a user-triggered build is more important as the user is waiting to
see the result of the build.